Refactor : 파ì파일 구조 변경
