+++ src/main/java/com/example/smartair/controller/ReissueController.java
@@ -1,7 +1,6 @@
 package com.example.smartair.controller;
 
-import com.example.smartair.entity.RefreshEntity;
-import com.example.smartair.entity.Role;
+import com.example.smartair.entity.login.RefreshEntity;
 import com.example.smartair.jwt.JWTUtil;
 import com.example.smartair.repository.RefreshRepository;
 import io.jsonwebtoken.ExpiredJwtException;
