+++ src/main/java/com/example/smartair/entity/login/CustomUserDetails.java
@@ -1,7 +1,6 @@
-package com.example.smartair.entity;
+package com.example.smartair.entity.login;
 
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.ArrayList;
