+++ src/main/java/com/example/smartair/entity/login/User.java
@@ -1,5 +1,7 @@
-package com.example.smartair.entity;
+package com.example.smartair.entity.login;
 
+import com.example.smartair.entity.BaseTimeEntity;
+import com.example.smartair.entity.login.Role;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -8,7 +10,7 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class User extends BaseTimeEntity{
+public class User extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
