+++ src/main/java/com/example/smartair/jwt/JWTUtil.java
@@ -1,8 +1,6 @@
 package com.example.smartair.jwt;
 
-import com.example.smartair.entity.Role;
 import io.jsonwebtoken.Jwts;
-import jakarta.servlet.http.Cookie;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
